blazingfast-api/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
httpx==0.25.0
requests==2.31.0
python-multipart==0.0.6
qdrant-client==1.7.0
numpy>=1.20.0
openai>=1.0.0